Some refactoring of mushishi, util, admin and various fixes
Cleaned up README: in progress
cleaned up term info: TODO: change to logging module
TODO: Use channel history instead of buffers where possible
added dm check to reaction
made admin.ld()'s reload functionality more robust TODO: more testing
and fixes
moved all plugin code to admin.py
cleaded up mushishi.py
refactored to remove close Admin.teardown() and mushishi.run()

+++ plugins/reaction.py
@@ -16,7 +16,7 @@
 from discord.ext import commands
 from discord.errors import NotFound, Forbidden
 from discord.ext.commands import Cog
-
+from discord import DMChannel
 from traceback import print_exc
 
 
@@ -61,17 +61,21 @@ class Reaction(Cog):
             # Erase message
             if r.emoji == "♻":
                 try:
+                    if isinstance(r.message.channel, DMChannel):
+                        return
+
                     rperc = r.count > len(r.message.channel.members) / 3
                     if rperc or user.id == self.bot.owner_id:
                         await r.message.delete()
                 except AttributeError:
                     await r.message.delete()
-                except Forbidden:  # fix by checking if channel is DM
-                    pass
+
             # Rerun command
             elif r.emoji == "🗜" and user.id == r.message.author.id:
                 try:
-                    await r.message.clear_reactions()
+                    for r in r.message.reactions:
+                        if r.me:
+                            await r.message.remove_reaction(r, self.bot.user)
                 finally:
                     await self.bot.process_commands(r.message)
